refactor: move constants to separate file
