+++ src/App.tsx
@@ -3,13 +3,6 @@ import { Root, rootLoader } from './routes/Root';
 import { ResultDetails, detailsLoader } from './routes/ResultDetails';
 import ErrorPage from './routes/ErrorPage';
 
-export const APP_CONFIG = {
-  DEFAULT_PAGE_SIZE: 10,
-  PAGE_SIZES: [10, 15, 20],
-  LOCAL_STORAGE_PREFIX: 'nasa-search-query',
-  SEARCH_BAR_PLACEHOLDER: 'search keywords',
-};
-
 const router = createBrowserRouter([
   {
     path: '/',
