+++ src/constants.ts
@@ -1,0 +1,6 @@
+export const APP_CONFIG = {
+  DEFAULT_PAGE_SIZE: 10,
+  PAGE_SIZES: [10, 15, 20],
+  LOCAL_STORAGE_PREFIX: 'nasa-search-query',
+  SEARCH_BAR_PLACEHOLDER: 'search keywords',
+};
